requirements.txt
# Core ML
torch==2.2.2
torchvision==0.17.2

# API
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Image processing
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.4

# Utilities
tqdm==4.66.2
python-multipart==0.0.9
pydantic==2.6.4